chore: refact router api to v2
